[TM Planner]: Show same-family supports
This lets us see which supports actually CAN support the unit, but only
can't be used due to already present units in the team. This helps in
teambuilding as sometimes we would want to change the team members to
use the support. This also shows the description of the currently
selected support, thus there is no more need to remove the unit to see
its support description.

`rowCallback` is called for every shown row, every redraw.